FIX : 다른 페이지에서 커뮤니티 화면이 출력되는 로직 수정
